Started moving auth routes to new register. sorted out new router much tidier now.
